+++ demos/main/src/main/java/com/google/android/exoplayer2/demo/EventLogger.java
@@ -528,6 +528,8 @@ import java.util.Locale;
         return "SEEK";
       case Player.DISCONTINUITY_REASON_SEEK_ADJUSTMENT:
         return "SEEK_ADJUSTMENT";
+      case Player.DISCONTINUITY_REASON_AD_INSERTION:
+        return "AD_INSERTION";
       case Player.DISCONTINUITY_REASON_INTERNAL:
         return "INTERNAL";
       default: